refactor(ui): extract MenuMixin for the gear-button app menu
Move build_menu, setup_donate_button, on_change_insight_calculation
and on_change_health_visualization into l5r/ui/menu.py. The mixin
still relies on self.widgets / self.tabs / self.ic_idx / self.sink1..4
plus self.please_donate from L5RCMCore.

Drops the lingering commented-out reset_geometry connect line while in
the area.

main.py: 1979 -> 1755 lines.
